move files around to support docker build and run
