+++ migrator/database/tests/mysql/test_check_migration.sql
